Current score type double

+++ src/application/AttemptController.java
@@ -38,7 +38,7 @@ public class AttemptController extends QuizController implements Initializable{
 		setWordAttempt((getWordAttempt()+1));
 		attemptNum.setText("attempt "+Integer.toString(getWordAttempt())+" of 2");
 		wordProgress.setText("play word "+Integer.toString(getWordProgress())+" of "+Integer.toString(getMaxNumWords()));
-		score.setText("current score: "+Integer.toString(getCurrentScore())); // TO UPDATE!
+		score.setText("current score: "+Double.toString(getCurrentScore())); // TO UPDATE!
 		
 		timer.setText("timer"); // TO DO!
